Move panel common code to detail implementation
